feat: add svdb output
This commit also features a fix of the config validation where the
schema had been indented one step too far a bit into the file resulting
in some validation simply not happening. This was the reason for quite a
few headaches that would have been solved with a working validation
process.