feat: add CNV HTML report (#40)
* Harmonize output file spec

In both Twist Solid and in the reports module, the output file
specification is called `output` instead of `output_files`. This commit
addresses this by applying the same convention for this pipeline.

* Add CNV filtering

These filters have been copied straight from the Twist Solid pipeline,
and that's a good start before we can start validating this pipeline
properly.

* Add CNV report to output

* Add simple CNV annotation and enable CNV report table

* Remove background filtering for now

The Twist_AF field should be added by SVDB query, but at this time we
don't have a panel of normals for this purpose. This will be added at a
later stage.

* Remove FLT3 for now

* Add gene annotations to the plots

* Fix issue with dry-run