Update the reference module version
The issue that I previously had has now been addressed, and there are no
releasese of this module avaliable, so I'll go with the commit ID. The
reason for not going with the branch name is that that messes with the
caching. If the branch in question is updated and there is an existing
cache pointing to that branch, the workflow will run with the cached
version. Using the commit ID gets around this issue.